+++ app/gui/src/project-view/components/GraphEditor/GraphNode/nodeVisualization.ts
@@ -7,7 +7,7 @@ import { type VisualizationDataSource } from '@/stores/visualization'
 import { type Opt } from '@/util/data/opt'
 import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
-import { computed, ref, toValue, watch } from 'vue'
+import { computed, ref, shallowRef, toValue, watch } from 'vue'
 import type { ComponentProps } from 'vue-component-type-helpers'
 import type { VisualizationIdentifier, VisualizationMetadata } from 'ydoc-shared/yjsModel'
 
@@ -22,30 +22,26 @@ interface Emit {
 interface NodeVisualizationOptions {
   vis: ToValue<Opt<VisualizationMetadata>>
   nodeHovered: ToValue<boolean>
-  nodeWidgetsSize: ToValue<Vec2>
-  nodePos: ToValue<Vec2>
+  nodeRect: ToValue<Rect>
   scale: ToValue<number>
   isFocused: ToValue<boolean>
   typeinfo: ToValue<Opt<TypeInfo>>
   dataSource: ToValue<Opt<VisualizationDataSource | RawDataSource>>
   hidden: ToValue<boolean>
   emit: Emit
-  onResize?: (rect0: Rect, rect1: Rect) => void
 }
 
 /** Composable managing the state of the visualization for a node. */
 export function useNodeVisualization({
   vis,
   nodeHovered,
-  nodeWidgetsSize,
-  nodePos,
+  nodeRect,
   scale,
   isFocused,
   typeinfo,
   dataSource,
   hidden,
   emit,
-  onResize,
 }: NodeVisualizationOptions) {
   const keyboard = injectBubblingKeyboard()
   const metadata = computed(() => toValue(vis))
@@ -95,40 +91,24 @@ export function useNodeVisualization({
     if (!visible && visualizationHovered.value) visualizationHovered.value = false
   })
 
-  const effectiveHeight = ref<number>()
-  const visibleVisHeight = computed((): number => {
-    if (!isVisualizationVisible.value || toValue(hidden)) return 0
-    return effectiveHeight.value ?? 0
-  })
-  const effectiveWidth = ref<number>()
-  const visibleVisWidth = computed((): number => {
-    if (!isVisualizationVisible.value || toValue(hidden)) return 0
-    return effectiveWidth.value ?? 0
-  })
-  const visibleSize = computed<Vec2 | undefined>((prev) => {
-    if (effectiveHeight.value == null || effectiveWidth.value == null) return
-    const size = new Vec2(visibleVisWidth.value, visibleVisHeight.value)
-    return prev?.equals(size) ? prev : size
-  })
-  let resizing = false
-  watch(visibleSize, (size1, size0) => {
-    if (!resizing || !size1 || !size0 || size1.equals(size0)) return
-    const widgets = toValue(nodeWidgetsSize)
-    const pos = toValue(nodePos)
-    const fullSize = (vizSize: Vec2) =>
-      new Vec2(Math.max(widgets.x, vizSize.x), widgets.y + vizSize.y)
-    const rect0 = new Rect(pos, fullSize(size0))
-    const rect1 = new Rect(pos, fullSize(size1))
-    onResize?.(rect0, rect1)
+  const visSize = shallowRef<Vec2>()
+  const visibleVisRect = computed((): Opt<Rect> => {
+    if (!isVisualizationVisible.value || toValue(hidden) || !visSize.value) return null
+    const nodeRectValue = toValue(nodeRect)
+    return new Rect(
+      nodeRectValue.pos,
+      new Vec2(visSize.value.x, nodeRectValue.size.y + visSize.value.y),
+    )
   })
 
   const visualization = computed((): ComponentProps<typeof GraphVisualization> => {
+    const { size: nodeSize, pos: nodePosition } = toValue(nodeRect)
     return {
       show: isVisualizationVisible.value,
       width: visualizationWidth.value,
-      nodeSize: toValue(nodeWidgetsSize),
+      nodeSize,
       scale: toValue(scale),
-      nodePosition: toValue(nodePos),
+      nodePosition,
       currentType: metadata.value?.identifier,
       dataSource: toValue(dataSource) ?? undefined,
       typeinfo: toValue(typeinfo) ?? undefined,
@@ -138,21 +118,19 @@ export function useNodeVisualization({
       isFullscreenAllowed: true,
       isResizable: true,
       'onUpdate:hovered': (event) => (visualizationHovered.value = event),
-      'onUpdate:effectiveHeight': (event) => (effectiveHeight.value = event),
-      'onUpdate:effectiveWidth': (event) => (effectiveWidth.value = event),
+      'onUpdate:effectiveSize': (event) => (visSize.value = event),
       'onUpdate:id': (event) => emit('update:visualizationId', event),
       'onUpdate:enabled': (event) => emit('update:visualizationEnabled', event),
       'onUpdate:height': (event) => emit('update:visualizationHeight', event),
       'onUpdate:width': (event) => (visualizationWidth.value = event),
-      'onUpdate:resizing': (event) => (resizing = event),
     }
   })
 
   return {
-    visualizationWidth: visibleVisWidth,
+    visualizationWidth,
     isVisualizationEnabled,
     isVisualizationPreviewed,
-    vizHeight: visibleVisHeight,
+    visRect: visibleVisRect,
     visualization,
   }
 }